+++ setup.cfg
@@ -6,6 +6,12 @@ addopts =
     --cov=ewave
     --cov-report term-missing
 
+[coverage:run]
+source = ewave
+
+[coverage:report]
+show_missing = true
+
 [compile_catalog]
 directory = ewave/locale
 domain = clld
